+++ database/migrations/2024_03_29_124507_create_attendances_table.php
@@ -10,6 +10,9 @@ return new class () extends Migration {
      */
     public function up(): void
     {
+        if (Schema::hasTable('attendances')) {
+            return;
+        }
         Schema::create('attendances', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
